Added README to project
Move CoreData Stack methods in a separate class SZDataStore.
Owner of this DataStore is the AppDelegate and the ViewController and the DataSource get a weak reference on it.
All save operations are now handled in the DataStore.

Add support for moveRows in the TabelView
- add tableView:moveRowAtIndexPath:fromIndexPath:toIndexPath:
- removed canMoveRowAtIndexPath
- added moveRowAtIndexPath:fromIndexPath:toIndexPath to datasource

Added two attributes to dataModel for testing the ordering (num and name)

Change delete implementation for testing.